Structure examples according to documentation
- add some missing examples
- fix documentation links in examples
- change development status to stable
