+++ example/README
@@ -1,0 +1,3 @@
+This directory contains the example code as used in the documentation.
+Each example contains a link to the related documentation chapter.
+Refer to that documentation for usage.
