+++ example/configuration/order_dependencies/test_sort_dependency.py
@@ -1,0 +1,15 @@
+"""Test showing the behavior of the --order-dependencies option.
+See https://pytest-order.readthedocs.io/en/stable/configuration.html#order-dependencies
+"""
+
+import pytest
+
+
+@pytest.mark.dependency(depends=["test_b"])
+def test_a():
+    assert True
+
+
+@pytest.mark.dependency
+def test_b():
+    assert True
