+++ example/configuration/order_group_scope/index_ordering/test_module2.py
@@ -1,0 +1,14 @@
+"""One of the test files showing the use of the --order-group-scope option.
+See https://pytest-order.readthedocs.io/en/stable/configuration.html#order-group-scope
+"""
+
+import pytest
+
+
+@pytest.mark.order(1)
+def test1():
+    pass
+
+
+def test2():
+    pass
