+++ example/configuration/order_scope/test_module1.py
@@ -1,0 +1,15 @@
+"""One of the test files showing the use of the --order-scope option.
+See https://pytest-order.readthedocs.io/en/stable/configuration.html#order-scope
+"""
+
+import pytest
+
+
+@pytest.mark.order(2)
+def test2():
+    pass
+
+
+@pytest.mark.order(1)
+def test1():
+    pass
