--- example/order_scope_level/feature2/test_b.py
@@ -1,11 +1,0 @@
-import pytest
-
-
-@pytest.mark.order(4)
-def test_four():
-    pass
-
-
-@pytest.mark.order(3)
-def test_three():
-    pass
