--- example/test.py
@@ -1,11 +1,0 @@
-"""Reference test without ordering from the quick start chapter.
-See https://pytest-order.readthedocs.io/en/stable/#quickstart
-"""
-
-
-def test_foo():
-    assert True
-
-
-def test_bar():
-    assert True
