--- example/test_relative_class_marker.py
@@ -1,18 +1,0 @@
-import pytest
-
-
-@pytest.mark.order(after="Test2")
-class Test1:
-    def test_1(self):
-        assert True
-
-    def test_2(self):
-        assert True
-
-
-class Test2:
-    def test_1(self):
-        assert True
-
-    def test_2(self):
-        assert True
