+++ example/usage/order_relative_to_other_tests/test_between_modules/test_module_a.py
@@ -1,0 +1,11 @@
+"""One of the example files showing relative ordering between modules.
+See https://pytest-order.readthedocs.io/en/stable/usage.html#referencing-of-tests-in-other-classes-or-modules  # noqa: E501
+"""
+
+
+class TestA:
+    def test_a(self):
+        assert True
+
+    def test_b(self):
+        assert True
