+++ example/usage/order_relative_to_other_tests/test_between_modules/test_module_c/test_submodule.py
@@ -1,0 +1,11 @@
+"""One of the example files showing relative ordering between modules.
+See https://pytest-order.readthedocs.io/en/stable/usage.html#referencing-of-tests-in-other-classes-or-modules  # noqa: E501
+"""
+
+
+def test_1():
+    pass
+
+
+def test_2():
+    pass
