+++ example/usage/order_relative_to_other_tests/test_combined.py
@@ -1,0 +1,15 @@
+"""Shows how combined ordinal and relative sorting is handled.
+See https://pytest-order.readthedocs.io/en/stable/usage.html#combination-of-absolute-and-relative-ordering  # noqa: E501
+"""
+
+import pytest
+
+
+@pytest.mark.order(index=0, after="test_second")
+def test_first():
+    assert True
+
+
+@pytest.mark.order(1)
+def test_second():
+    assert True
