+++ setup.py
@@ -27,13 +27,13 @@ setup(
             "pytest_order = pytest_order.plugin",
         ]
     },
-    python_requires=">=3.6",
+    python_requires=">=3.7",
     install_requires=[
         "pytest>=5.0; python_version < '3.10'",
         "pytest>=6.2.4; python_version >= '3.10'",
     ],
     classifiers=[
-        "Development Status :: 4 - Beta",
+        "Development Status :: 5 - Production/Stable",
         "Intended Audience :: Developers",
         "License :: OSI Approved :: MIT License",
         "Operating System :: POSIX",
